refactor: convert OfflineScreen to reusable OfflineErrorComponent and add to DashboardScreen.kt
